feat: TokenExpiredError 사용해서 토큰이 만료 되면 클라이언트에게 메세지를 보내도록 로직 수정

+++ controller/likeController.js
@@ -1,10 +1,18 @@
 const conn = require("../db/mariadb");
 const { StatusCodes } = require("http-status-codes");
 const { decodedJWT } = require("../helper");
+const { TokenExpiredError } = require("jsonwebtoken");
 
 const addLike = async (req, res) => {
     const { id } = req.params;
-    const decoded = decodedJWT(req, res);
+    const decoded = decodedJWT(req);
+
+    if (decoded instanceof TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해 주세요."
+        })
+    }
+
 
     let sql = "INSERT INTO LIKES_TB (user_id, liked_book_id) VALUES (?, ?)";
 
@@ -19,7 +27,13 @@ const addLike = async (req, res) => {
 
 const removeLike = async (req, res) => {
     const { id } = req.params;
-    const decoded = decodedJWT(req, res);
+    const decoded = decodedJWT(req);
+
+    if (decoded instanceof TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해 주세요."
+        })
+    }
 
     let sql = "DELETE FROM LIKES_TB WHERE user_id = ? AND liked_book_id = ?";
